requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
alpaca-py>=0.42.0
pychalk>=2.0.1
statsmodels>=0.14.0
deap>=1.4.0
tsfracdiff>=1.0.4
stumpy>=1.13.0
quantstats>=0.0.64
